requirements.txt
# Core dependencies
asyncio-mqtt>=0.11.1
click>=8.1.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# GUI dependencies
PySide6>=6.5.0

# Database and ORM
alembic>=1.11.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0
mypy>=1.4.0

# Logging and configuration
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Optional: For future QQ bot integration
nonebot2>=2.0.0